EVAKA-4043: backup childen in contact report + fix file names
