+++ service/src/main/kotlin/fi/espoo/evaka/reports/FamilyConflictReport.kt
@@ -6,134 +6,98 @@ package fi.espoo.evaka.reports
 
 import fi.espoo.evaka.Audit
 import fi.espoo.evaka.shared.auth.AccessControlList
+import fi.espoo.evaka.shared.auth.AclAuthorization
 import fi.espoo.evaka.shared.auth.AuthenticatedUser
 import fi.espoo.evaka.shared.auth.UserRole
 import fi.espoo.evaka.shared.db.Database
-import fi.espoo.evaka.shared.db.mapColumn
-import fi.espoo.evaka.shared.db.mapNullableColumn
+import fi.espoo.evaka.shared.db.bindNullable
+import fi.espoo.evaka.shared.db.getUUID
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 import java.util.UUID
 
 @RestController
-class FamilyContactReportController(private val acl: AccessControlList) {
-    @GetMapping("/reports/family-contacts")
-    fun getFamilyContactsReport(
-        db: Database,
-        user: AuthenticatedUser,
-        @RequestParam unitId: UUID
-    ): ResponseEntity<List<FamilyContactReportRow>> {
-        Audit.FamilyContactReportRead.log()
-        acl.getRolesForUnit(user, unitId).requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.FINANCE_ADMIN, UserRole.DIRECTOR, UserRole.UNIT_SUPERVISOR, UserRole.SPECIAL_EDUCATION_TEACHER)
-        return db.read { it.getFamilyContacts(unitId) }.let { ResponseEntity.ok(it) }
+class FamilyConflictReportController(private val acl: AccessControlList) {
+    @GetMapping("/reports/family-conflicts")
+    fun getFamilyConflictsReport(db: Database, user: AuthenticatedUser): ResponseEntity<List<FamilyConflictReportRow>> {
+        Audit.FamilyConflictReportRead.log()
+        user.requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.FINANCE_ADMIN, UserRole.UNIT_SUPERVISOR)
+        return db.read { it.getFamilyConflicts(acl.getAuthorizedUnits(user)) }
+            .let { ResponseEntity.ok(it) }
     }
 }
 
-private fun Database.Read.getFamilyContacts(unitId: UUID): List<FamilyContactReportRow> {
+private fun Database.Read.getFamilyConflicts(authorizedUnits: AclAuthorization): List<FamilyConflictReportRow> {
     // language=sql
     val sql =
         """
+        WITH child_conflicts AS (
+            SELECT head_of_child as id, count(*) as child_conflict_count
+            FROM fridge_child
+            WHERE conflict = true
+            GROUP BY head_of_child
+        ), partner_conflicts AS (
+            SELECT person_id as id, count(*) as partner_conflict_count
+            FROM fridge_partner
+            WHERE conflict = true
+            GROUP BY person_id
+        ), conflicts AS (
             SELECT
-                ch.id,
-                ch.first_name,
-                ch.last_name,
-                ch.social_security_number,
-                ch.street_address,
-                ch.postal_code,
-                ch.post_office,
-                dg.name AS group_name,
-
-                hoc.id AS hoc_id,
-                hoc.first_name AS hoc_first_name,
-                hoc.last_name AS hoc_last_name,
-                hoc.phone AS hoc_phone,
-                hoc.email AS hoc_email,
-
-                gu1.id AS gu1_id,
-                gu1.first_name AS gu1_first_name,
-                gu1.last_name AS gu1_last_name,
-                gu1.phone AS gu1_phone,
-                gu1.email AS gu1_email,
-
-                gu2.id AS gu2_id,
-                gu2.first_name AS gu2_first_name,
-                gu2.last_name AS gu2_last_name,
-                gu2.phone AS gu2_phone,
-                gu2.email AS gu2_email
-            FROM placement pl
-            JOIN person ch ON ch.id = pl.child_id
-            LEFT JOIN daycare_group_placement dgp ON pl.id = dgp.daycare_placement_id AND daterange(dgp.start_date, dgp.end_date, '[]') @> current_date
-            LEFT JOIN daycare_group dg ON dg.id = dgp.daycare_group_id
-            LEFT JOIN fridge_child fc ON ch.id = fc.child_id AND daterange(fc.start_date, fc.end_date, '[]') @> current_date
-            LEFT JOIN person hoc ON hoc.id = fc.head_of_child
-            LEFT JOIN guardian g1 ON ch.id = g1.child_id AND g1.guardian_id = (SELECT min(guardian_id::text)::uuid FROM guardian where child_id = ch.id)
-            LEFT JOIN guardian g2 ON ch.id = g2.child_id AND g2.guardian_id = (SELECT max(guardian_id::text)::uuid FROM guardian where child_id = ch.id) AND g2.guardian_id != (SELECT min(guardian_id::text)::uuid FROM guardian where child_id = ch.id)
-            LEFT JOIN person gu1 ON gu1.id = g1.guardian_id
-            LEFT JOIN person gu2 ON gu2.id = g2.guardian_id
-            WHERE pl.unit_id = :unitId AND daterange(pl.start_date, pl.end_date, '[]') @> current_date
-            ORDER BY dg.name, ch.last_name, ch.first_name
+                p.id,
+                p.first_name,
+                p.last_name,
+                p.social_security_number,
+                coalesce(cc.child_conflict_count, 0) as child_conflict_count,
+                coalesce(pc.partner_conflict_count, 0) as partner_conflict_count
+            FROM child_conflicts cc
+            FULL JOIN partner_conflicts pc ON cc.id = pc.id
+            JOIN person p ON p.id = coalesce(cc.id, pc.id)
+        )
+        SELECT 
+            ca.name AS care_area_name,
+            u.id AS unit_id,
+            u.name AS unit_name,
+            co.id,
+            co.first_name,
+            co.last_name,
+            co.social_security_number,
+            co.child_conflict_count,
+            co.partner_conflict_count,
+            pu.unit_id
+        FROM conflicts co
+        JOIN primary_units_view pu ON co.id = pu.head_of_child
+        JOIN daycare u ON u.id = pu.unit_id
+        JOIN care_area ca ON ca.id = u.care_area_id
+        WHERE (:units::uuid[] IS NULL OR u.id = ANY(:units))
+        ORDER BY ca.name, u.name, co.last_name, co.first_name
         """.trimIndent()
-
     return createQuery(sql)
-        .bind("unitId", unitId)
-        .map { rs, ctx ->
-            FamilyContactReportRow(
-                id = ctx.mapColumn(rs, "id"),
-                firstName = rs.getString("first_name") ?: "",
-                lastName = rs.getString("last_name") ?: "",
-                ssn = rs.getString("social_security_number"),
-                streetAddress = rs.getString("street_address"),
-                postalCode = rs.getString("postal_code"),
-                postOffice = rs.getString("post_office"),
-                group = rs.getString("group_name"),
-                headOfChild = ctx.mapNullableColumn<UUID>(rs, "hoc_id")?.let {
-                    Contact(
-                        firstName = rs.getString("hoc_first_name") ?: "",
-                        lastName = rs.getString("hoc_last_name") ?: "",
-                        phone = rs.getString("hoc_phone") ?: "",
-                        email = rs.getString("hoc_email") ?: ""
-                    )
-                },
-                guardian1 = ctx.mapNullableColumn<UUID>(rs, "gu1_id")?.let {
-                    Contact(
-                        firstName = rs.getString("gu1_first_name") ?: "",
-                        lastName = rs.getString("gu1_last_name") ?: "",
-                        phone = rs.getString("gu1_phone") ?: "",
-                        email = rs.getString("gu1_email") ?: ""
-                    )
-                },
-                guardian2 = ctx.mapNullableColumn<UUID>(rs, "gu2_id")?.let {
-                    Contact(
-                        firstName = rs.getString("gu2_first_name") ?: "",
-                        lastName = rs.getString("gu2_last_name") ?: "",
-                        phone = rs.getString("gu2_phone") ?: "",
-                        email = rs.getString("gu2_email") ?: ""
-                    )
-                }
+        .bindNullable("units", if (authorizedUnits != AclAuthorization.All) authorizedUnits.ids else null)
+        .map { rs, _ ->
+            FamilyConflictReportRow(
+                careAreaName = rs.getString("care_area_name"),
+                unitId = rs.getUUID("unit_id"),
+                unitName = rs.getString("unit_name"),
+                id = rs.getUUID("id"),
+                firstName = rs.getString("first_name"),
+                lastName = rs.getString("last_name"),
+                socialSecurityNumber = rs.getString("social_security_number"),
+                partnerConflictCount = rs.getInt("partner_conflict_count"),
+                childConflictCount = rs.getInt("child_conflict_count")
             )
         }
-        .list()
+        .toList()
 }
 
-data class FamilyContactReportRow(
+data class FamilyConflictReportRow(
+    val careAreaName: String,
+    val unitId: UUID,
+    val unitName: String,
     val id: UUID,
-    val firstName: String,
-    val lastName: String,
-    val ssn: String?,
-    val group: String?,
-    val streetAddress: String,
-    val postalCode: String,
-    val postOffice: String,
-    val headOfChild: Contact?,
-    val guardian1: Contact?,
-    val guardian2: Contact?
-)
-
-data class Contact(
-    val firstName: String,
-    val lastName: String,
-    val phone: String,
-    val email: String
+    val firstName: String?,
+    val lastName: String?,
+    val socialSecurityNumber: String?,
+    val partnerConflictCount: Int,
+    val childConflictCount: Int
 )
